refactor(game): change file name for consistency
