+++ service/app/src/entities/game/api/deleteGame.ts
