+++ service/app/src/entities/game/api/playGame.ts
