+++ service/app/src/entities/game/api/shareGame.ts
